requirements.txt
streamlit
pandas
sqlalchemy
plotly
google-genai
groq
pdfplumber
python-dotenv
PyMuPDF
